Port StopLine and StopArea controllers to ROS. Add these and FollowLane to build list. Delete LaneHeading controller, replaced by course->desired_heading() (FollowLane controller not yet invoked by Run controller.)

+++ stacks/art_vehicle/art_nav/src/navigator/stop_area.h
@@ -24,7 +24,7 @@ public:
 
 private:
   // .cfg variables
-  float stop_approach_speed;		// speed while approaching stop
+  double stop_approach_speed;          // speed while approaching stop
 
   // controller state
   bool in_safety_area;
